[EC-112] Moved OmmersValidator to new validators package
