+++ src/main/scala/io/iohk/ethereum/validators/OmmersValidator.scala
@@ -1,7 +1,8 @@
-package io.iohk.ethereum.network.p2p.validators
+package io.iohk.ethereum.validators
 
 import io.iohk.ethereum.domain.{BlockHeader, Blockchain}
-import io.iohk.ethereum.network.p2p.validators.OmmersValidator.OmmersError._
+import io.iohk.ethereum.network.p2p.validators.BlockHeaderValidator
+import io.iohk.ethereum.validators.OmmersValidator.OmmersError._
 
 object OmmersValidator {
